Add operator to Porter docs
